Add legacy classes to separate module and enable target scaling
