+++ src/gnn_reco/components/loss_functions.py
@@ -216,41 +216,3 @@ class VonMisesFisher2DLoss(VonMisesFisherLoss):
         ], dim=1)
 
         return self._evaluate(p, t)
-
-
-class LegacyVonMisesFisherLoss(LossFunction):
-    def _forward(self, prediction: Tensor, target: Tensor) -> Tensor:
-        """Original function for computing von Mises-Fisher loss.
-
-        Repesents a single angle (graph[target]) as a 3D vector (sine(angle),
-        cosine(angle), 1) and calculates the 3D VonMisesFisher loss of the
-        angular difference between the 3D vector representations.
-
-        Args:
-            prediction (Tensor): Output of the model. Must have shape [N, 3]
-                where 0th column is a prediction of sine(angle) and 1st column
-                is prediction of cosine(angle) and 2nd column is an estimate of
-                kappa.
-            target (Tensor): Target tensor, extracted from graph object.
-
-        Returns:
-            loss (Tensor): Elementwise von Mises-Fisher loss terms. Shape [N,]
-        """
-        k = torch.abs(prediction[:,2])
-        angle  = target  # Original: graph[target].squeeze(1)
-        u_1 = (1/torch.sqrt(torch.tensor(2,dtype = torch.float)))*torch.sin(angle)
-        u_2 = (1/torch.sqrt(torch.tensor(2,dtype = torch.float)))*torch.cos(angle)
-        u_3 = (1/torch.sqrt(torch.tensor(2,dtype = torch.float)))*(1)
-
-        norm_x  = torch.sqrt(1 + prediction[:,0]**2 + prediction[:,1]**2)
-
-        x_1 = (1/norm_x)*prediction[:,0]
-        x_2 = (1/norm_x)*prediction[:,1]
-        x_3 = (1/norm_x)*(1)
-
-        dotprod = u_1*x_1 + u_2*x_2 + u_3*x_3
-        logc_3 = - torch.log(k) + k + torch.log(1 - torch.exp(-2*k))
-        elements = -k*dotprod + logc_3
-
-        return elements
-
